feat(utils): add `del()` function to delete properties
The changes in this commit add a new `del()` function to the `utils` namespace in the `src/index.ts` file. This function allows deleting properties from an object based on a given path array.

The changes also update the `createEventsWritable()` function to handle the new "delete" event type, which is represented by the "-" character in the serialized event string.

Additionally, the `MetadataDelete` type is introduced to represent the metadata for a delete event, and the `Metadata` type is updated to be a union of `MetadataUpdate` and `MetadataDelete`.

The new `del()` function and the corresponding changes in the `createEventsWritable()` function are important to provide a complete set of operations for working with the events writable stream.